Adjust ui size and position
